+++ azplugins/__init__.py
@@ -1,0 +1,1 @@
+from hoomd.azplugins import pair