+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 [TOC]
 
+---------
 ## v0.1.0
 
 Not yet released
+
+*New features*
+
+* A framework is configured for developing pair potentials.
+* `pair.ashbaugh` implements the Ashbaugh-Hatch (Lennard-Jones perturbation)
+   potential.